Revert "feat: create importer online component (#1541)"
This reverts commit 64daad0ca0034513af11d6d9207dfe8fe779e1bf.
