+++ Cargo.toml
@@ -156,6 +156,14 @@ path = "src/bin/rpc_downloader.rs"
 name = "importer-offline"
 path = "src/bin/importer_offline.rs"
 
+[[bin]]
+name = "importer-online"
+path = "src/bin/importer_online.rs"
+
+[[bin]]
+name = "run-with-importer"
+path = "src/bin/run_with_importer.rs"
+
 # ------------------------------------------------------------------------------
 # Features
 # ------------------------------------------------------------------------------
