+++ e2e/cloudwalk-contracts/integration/test/helpers/rpc.ts
@@ -44,7 +44,7 @@ export let ETHERJS = new JsonRpcProvider(providerUrl, undefined);
 
 export function updateProviderUrl(providerName: string) {
     switch (providerName) {
-        case "importer":
+        case "run-with-importer":
             providerUrl = "http://localhost:3001?app=e2e";
             break;
         case "stratus":
