+++ src/eth/consensus/log_entry.rs
