--- src/eth/importer/mod.rs
@@ -1,6 +1,0 @@
-#[allow(clippy::module_inception)]
-mod importer;
-mod importer_config;
-
-pub use importer::Importer;
-pub use importer_config::ImporterConfig;
