+++ src/eth/mod.rs
@@ -1,7 +1,6 @@
 pub mod codegen;
 pub mod consensus;
 pub mod executor;
-pub mod importer;
 pub mod miner;
 pub mod primitives;
 pub mod rpc;
