+++ src/main.rs
@@ -1,10 +1,8 @@
 use std::sync::Arc;
 
 use stratus::config::StratusConfig;
-use stratus::eth::consensus::simple_consensus::SimpleConsensus;
-use stratus::eth::consensus::Consensus;
 use stratus::eth::rpc::serve_rpc;
-use stratus::infra::BlockchainClient;
+use stratus::eth::Consensus;
 use stratus::GlobalServices;
 
 fn main() -> anyhow::Result<()> {
@@ -13,58 +11,28 @@ fn main() -> anyhow::Result<()> {
 }
 
 async fn run(config: StratusConfig) -> anyhow::Result<()> {
-    // Init services
+    // init services
     let storage = config.storage.init()?;
-
-    // Init miner
-    let miner = if config.follower {
-        config.miner.init_external_mode(Arc::clone(&storage))?
-    } else {
-        config.miner.init(Arc::clone(&storage))?
-    };
-
-    // Init executor
+    let miner = config.miner.init(Arc::clone(&storage))?;
     let executor = config.executor.init(Arc::clone(&storage), Arc::clone(&miner));
+    let consensus = Consensus::new(
+        Arc::clone(&storage),
+        Arc::clone(&miner),
+        config.storage.perm_storage.rocks_path_prefix.clone(),
+        config.clone().candidate_peers.clone(),
+        None,
+        config.rpc_server.address,
+        config.grpc_server_address,
+    ); // for now, we force None to initiate with the current node being the leader
 
-    // Init chain
-    let chain = if config.follower {
-        let importer_config = config.importer.as_ref().ok_or(anyhow::anyhow!("importer config is not set"))?;
-        Some(Arc::new(
-            BlockchainClient::new_http_ws(
-                importer_config.external_rpc.as_ref(),
-                importer_config.external_rpc_ws.as_deref(),
-                importer_config.external_rpc_timeout,
-            )
-            .await?,
-        ))
-    } else {
-        None
-    };
-
-    // Init consensus
-    let consensus: Arc<dyn Consensus> = Arc::new(SimpleConsensus::new(Arc::clone(&storage), chain.clone()));
-
-    // Init importer
-    if config.follower {
-        if let Some(importer_config) = &config.importer {
-            if let Some(chain) = chain {
-                importer_config.init(Arc::clone(&executor), Arc::clone(&miner), Arc::clone(&storage), chain)?;
-            } else {
-                return Err(anyhow::anyhow!("chain is not initialized"));
-            }
-        } else {
-            return Err(anyhow::anyhow!("importer config is not set"));
-        }
-    }
-
-    // Init RPC server
+    // start rpc server
     serve_rpc(
-        // Services
+        // services
         Arc::clone(&storage),
         executor,
         miner,
         consensus,
-        // Config
+        // config
         config.clone(),
         config.rpc_server,
         config.executor.executor_chain_id.into(),
